modified:   Abstract/Resampler.py 	modified:   Implementations/perturbers/perturbers.py 	modified:   Implementations/resamplers/resamplers.py 	modified:   Implementations/solvers/DeterministicSolvers.py 	modified:   Implementations/solvers/StochasticSolvers.py 	new file:   datasets/FLU_HOSPITALIZATIONS.csv 	modified:   main.py 	new file:   scratch.ipynb

+++ Implementations/solvers/DeterministicSolvers.py
@@ -16,7 +16,7 @@ class EulerSolver(Integrator):
         for particle in particleArray: 
             particle.observation = 0
 
-        for _ in range(1/dt): 
+        for _ in range(int(1/dt)): 
 
             for j,particle in enumerate(particleArray): 
                 '''This loop runs over the particleArray, performing the integration in RHS for each one'''
